+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Removed sandbox mode from macOS
+- macOS Script:
+  - Install Flutter SDK
+  - Unzip Flutter SDK
+  - Add Flutter SDK to PATH
+  - Install Git (requires `brew`)
+  - Install Android Studio
+  - Install IntelliJ IDEA
+  - Install Visual Studio Code
+
 ## [0.0.2] - 2020-07-23
 
 ### Added